+++ src/main/java/lotto/domain/lottofactory/LottoConstant.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottofactory;
 
 public class LottoConstant {
     public static final int LIMIT_MINIMUM_NUMBER = 1;
