+++ src/main/java/lotto/domain/shufflerule/FixShuffle.java
@@ -1,15 +1,13 @@
-package lotto.domain.ShuffleRule;
+package lotto.domain.shufflerule;
 
-import lotto.domain.LottoConstant;
-import lotto.domain.LottoNumber;
+import lotto.domain.lottofactory.LottoConstant;
+import lotto.domain.lottofactory.LottoNumber;
 
-import java.util.Collections;
 import java.util.List;
 
-public class RandomShuffle implements Shuffle {
+public class FixShuffle implements Shuffle {
     @Override
     public List<LottoNumber> getShuffledLottoNumbers(List<LottoNumber> lottoNumbers) {
-        Collections.shuffle(lottoNumbers);
         return lottoNumbers.subList(LottoConstant.SUBLIST_FIRST_INDEX, LottoConstant.LOTTO_NUMBER_SIZE);
     }
 }
