+++ src/main/java/lotto/domain/shufflerule/RandomShuffle.java
@@ -1,0 +1,15 @@
+package lotto.domain.shufflerule;
+
+import lotto.domain.lottofactory.LottoConstant;
+import lotto.domain.lottofactory.LottoNumber;
+
+import java.util.Collections;
+import java.util.List;
+
+public class RandomShuffle implements Shuffle {
+    @Override
+    public List<LottoNumber> getShuffledLottoNumbers(List<LottoNumber> lottoNumbers) {
+        Collections.shuffle(lottoNumbers);
+        return lottoNumbers.subList(LottoConstant.SUBLIST_FIRST_INDEX, LottoConstant.LOTTO_NUMBER_SIZE);
+    }
+}
