+++ src/test/java/lotto/domain/lottofactory/LottoAmountTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottofactory;
 
 import org.junit.jupiter.api.Test;
